examples/quickstart/requirements.txt
pgai
psycopg
psycopg-pool
pgvector
python-dotenv
openai
datasets
numpy
structlog